requirements.txt
streamlit
crewai
langchain-google-genai
yfinance
python-dotenv
google-generativeai